refactor(carousel): extract useCarousel hook and cache media query
